requirements.txt
Flask==2.3.3
Pillow==10.0.0
stepic==0.5
cryptography==41.0.3